Add a functional decimal point button
The decimal point button works as expected and it
can only be used once in each number of the
operation.

The "equal" button is only enabled when we have
both numbers and the operation registered.

All the other buttons, like CE and AC, as well as
the operators, are enabled when it makes sense to
use them.